dashboard/requirements.txt
flask==2.3.2
requests==2.31.0
mysql-connector-python==8.1.0
